requirements.txt
fpdf==1.7.2
Pillow==9.4.0
